moving xy component as composite-y to avoid confusing with xyChartView
