+++ examples/js/app/example1.js
@@ -28,7 +28,7 @@ define([
       config: {
         bindings: [
           {
-            sourceComponent: 'xyChart',
+            sourceComponent: 'compositeY',
             sourceModel: 'config',
             sourcePath: 'plot',
             targetComponent: 'controlPanel',
@@ -39,7 +39,7 @@ define([
       }
     }],
     components: [{
-      type: 'xyChart',
+      type: 'compositeY',
       config: {
         el: '#complexChart-xyChart',
         marginInner: 10,
@@ -198,7 +198,7 @@ define([
   simpleChartView.setData(simpleData)
   simpleChartView.setConfig({
     components: [{
-      type: 'xyChart',
+      type: 'compositeY',
       config: {
         el: '#simpleChart',
         plot: {
