+++ src/js/contrail-charts/components/composite-y/CompositeYChartConfigModel.js
