+++ src/js/contrail-charts/components/composite-y/ScatterBubbleChartView.js
@@ -3,7 +3,7 @@
  */
 var _ = require('lodash')
 var d3 = require('d3')
-var XYChartSubView = require('contrail-charts/components/xy/XYChartSubView')
+var XYChartSubView = require('contrail-charts/components/composite-y/XYChartSubView')
 
 var ScatterBubbleChartView = XYChartSubView.extend({
   tagName: 'div',
