+++ src/js/contrail-charts/components/navigation/NavigationView.js
@@ -7,7 +7,7 @@ var d3 = require('d3')
 var Events = require('contrail-charts-events')
 var ContrailChartsView = require('contrail-charts-view')
 var DataProvider = require('contrail-charts/handlers/DataProvider')
-var CompositeYChartView = require('contrail-charts/components/xy/CompositeYChartView')
+var CompositeYChartView = require('contrail-charts/components/composite-y/CompositeYChartView')
 
 var NavigationView = ContrailChartsView.extend({
   tagName: 'div',
